Fixed Travis CI failure
Turns out that the fix was more convoluted than anticipated,
and it also involved disabling sbt forking the tests.

This revolves around the system properties not being passed
to the forked processes (see [0]) and also the fact that the
logic in the code would try to access default values that happened
to be correct locally, but wrong remotely.

Lessons learned: sample values in configuration files should always
point to nonsense values, so that we don't get "success by mistake."

[0] http://stackoverflow.com/questions/10155870/setting-system-properties-with-sbt-run