[Finishes #111282636] Configuration changes.
The configuration files are picked from the CLASSPATH and would be
'frozen' in the JAR when the target is built via the `sbt docker`
plugin.

In order to make them editable and the changes picked up at server
restart, we copy them in the /etc/sentinel folder of the container
and add that folder to the CLASSPATH (by fudging the ./bin/sentinel
script).

This is an irritating quirk of the `sbt docker` plugin, which should
be fixed in a more portable way eventually.