[Starts #109376462] Automates creation of full Service
Using Docker Machine we can now build and run containers
from both Linux and OS X development environments.

Also automated the creation of the API Server container,
refined the script to fix the flawed shell script.

There is still work necessary to complete the script that
runs the service.

Finally, some modification may be necessary to launch this
on an AWS EC2 instance.