Fixes issues around DB initialization in tests.
[Finishes #111291602]

Among other things, during app initialization, the test suite(s) were
unable to inject the test db URI, so the one defined in application.conf
was always picked.

This was contrary to the current approach of defining the server hostname/IP
during the test run (see bin/run_tests.sh) and causes issues also in the
Travis CI environment.

We have currently fixed this with a workaround, but we would like to find
a more stable solution.

A number of other fixes/refactorings done around tests.